Everything works

+++ main.cpp
@@ -19,26 +19,16 @@ int main() {
     std::ifstream ans("ans.txt");
     std::ofstream output("output.txt");
 
-    //AbstractMatrix::set_multiplier();
-
-    SquareMatrix<long long > matrix1(100);
+    SquareMatrix<unsigned > matrix1(100);
     matrix1.set_thread_cnt(8);
     matrix1.read(first);
-    //std::cout << '\n';
-    //matrix1.set_multiplier(multiplier<int, Buffer<int>>)
-    SquareMatrix<long long > matrix2(100);
+    SquareMatrix<unsigned > matrix2(100);
     matrix2.read(second);
 
-    SquareMatrix<long long > res(100);
+    SquareMatrix<unsigned > res(100);
     res.read(ans);
-    //std::cout <<'\n';
 
-   // SquareMatrix<int> matrix3(2);
-
-
-
-
-    SquareMatrix<long long > matrix3 = matrix1 * matrix2;
+    SquareMatrix<unsigned > matrix3 = matrix1 * matrix2;
     std::cout << (res == matrix3)<< std::endl;
